+++ lib/simple_state.dart
@@ -2,6 +2,7 @@ library simple_state;
 
 export 'src/reaction.dart';
 export 'src/observer.dart';
+export 'src/extensions.dart';
 export 'src/observable/observable.dart';
 export 'src/observable/observable_map.dart';
 export 'src/observable/observable_set.dart';