+++ lib/src/observable/observable_set.dart
@@ -2,7 +2,7 @@ import 'dart:collection';
 import 'package:flutter/foundation.dart';
 
 class ObservableSet<T> extends ChangeNotifier with SetMixin<T> {
-  ObservableSet(Set<T> value) : _value = value;
+  ObservableSet([Set<T>? value]) : _value = value ?? {};
 
   final Set<T> _value;