+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+## [1.1.0]
+
+- Refinement of examples.
+- Added observable "when" extensions.
+- Fix ObservableList set length.
+- Implemented default constructors for Observable collections.
+- Renamed ObserverListenerMixin.
+
 ## [1.0.1]
 
-- change after pub.dev analysis
+- Change after pub.dev analysis.
 
 ## [1.0.0]
 
-- initial release.
+- Initial release.